# public/openapi.yaml
openapi: 3.1.0
info:
  title: GPTAdmin MCP Relay
  version: "1.0.0"
  description: |
    Compact control API: discover → schema → execute. Poll job when background=true.

    Shell hosts and MCP services are exposed as GPTAdmin servers with ids like shell:<server_name>.
    The hub itself is exposed as target "hub" for registry and approval tools.

    Network Tunnel v1 exposes bounded control operations only. Raw proxy streams, offers and
    SOCKS/HTTP data endpoints remain separate; see docs/NETWORK_PROXY.md and
    docs/NETWORK_PROXY_THREAT_MODEL.md.
servers:
  - url: https://u-f1102930.t.gptadmin.bezrabotnyi.com
security:
  - bearerAuth: []
paths:
...
  /webhook-routes:
    get:
      operationId: listWebhookRoutes
      summary: List webhook route metadata
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Route metadata without secrets
    post:
      operationId: createWebhookRoute
      summary: Create an operator-owned webhook route
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRoute"
      responses:
        "201":
          description: Created route metadata without secrets
  /webhook-routes/{route}:
    put:
      operationId: replaceWebhookRoute
      summary: Replace an operator-owned webhook route
      security:
        - bearerAuth: []
      parameters:
        - name: route
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRoute"
      responses:
        "200":
          description: Updated route metadata without secrets
    delete:
      operationId: deleteWebhookRoute
      summary: Delete an operator-owned webhook route
      security:
        - bearerAuth: []
      parameters:
        - name: route
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Route deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    webhookToken:
      type: http
      scheme: bearer
  schemas:
...
    WebhookRoute:
      type: object
      additionalProperties: false
      required: [id, action]
      properties:
        id:
          type: string
        token:
          type: string
          writeOnly: true
        hmac_secret:
          type: string
          writeOnly: true
        signature_version:
          type: string
          enum: [v1, v2]
        max_skew_seconds:
          type: integer
          minimum: 1
        action:
          type: object
          additionalProperties: false
          required: [kind, target]
          properties:
            kind:
              type: string
              enum: [mcp, prompt, shell]
            target:
              type: string
            approval_mode:
              type: string
              enum: [ask_before_write, bounded_autonomous]
            tool:
              type: string
            arguments:
              type: object
              additionalProperties: true
            prompt:
              type: string
            prompt_arg:
              type: string
            command:
              type: string
            cwd:
              type: string
        callback:
          type: object
          additionalProperties: false
          required: [url]
          properties:
            url:
              type: string
            token:
              type: string
              writeOnly: true
            hmac_secret:
              type: string
              writeOnly: true
...